requirements.txt
streamlit
torch~=2.2.0
st-annotated-text
streamlit-extras
numpy~=1.26.4